requirements.txt
icrawler
Pillow
